Unify TestViewModel and ProgressViewModel, add loading indicator with delay for last question
